NetworkingScreen with tab navigation
