+++ js/components/developer/networking-screen/index.js
@@ -1,23 +1,9 @@
 import React, { Component } from 'react';
-import { View, Text, Image, StyleSheet, ListView } from 'react-native';
-import { Container, Content, Spinner } from 'native-base';
+import { View, Text } from 'react-native';
+import { Tab, Tabs  } from 'native-base';
+import MovieExample from './movieExample';
 
-const REQUEST_URL = 'https://raw.githubusercontent.com/facebook/react-native/master/docs/MoviesExample.json';
-
-// Component based on react-native 0.19 tutorial
-// http://facebook.github.io/react-native/releases/0.19/docs/tutorial.html
-export default class MovieExample extends Component {
-
-  // Constructor setting intitial state
-  constructor(props) {
-    super(props);
-    this.state = {
-      dataSource: new ListView.DataSource({
-        rowHasChanged: (row1, row2) => row1 !== row2,
-      }),
-      loaded: false,
-    };
-  }
+export default class NetworkingScreen extends Component {
 
   // Navigation information
   static navigationOptions = {
@@ -27,97 +13,17 @@ export default class MovieExample extends Component {
 		title: 'Networking Screen'
 	};
 
-  // Called once component has been mounted
-  componentDidMount() {
-    // Setting a delay so there is always time to see the loading screen. Do NOT do this!
-    setTimeout(
-      () => {
-        this.fetchData();
-      },
-      1000
-    );
-  }
-
-  // Fetch data from REQUEST_URL and update state
-  fetchData() {
-    fetch(REQUEST_URL)
-      .then((response) => response.json())
-      .then((responseData) => {
-        this.setState({
-          dataSource: this.state.dataSource.cloneWithRows(responseData.movies),
-          loaded: true,
-        });
-      })
-      .done();
-  }
-
   // Render the component
   render() {
-    if (!this.state.loaded) {
-      return this.renderLoadingView();
-    }
-    return (
-      <ListView
-        dataSource={this.state.dataSource}
-        renderRow={this.renderMovie}
-        style={styles.listView}
-      />
-    );
-  }
-
-  // Render loading screen
-  renderLoadingView() {
-    return (
-      <Container>
-        <Content>
-          <Spinner color='blue' />
-        </Content>
-      </Container>
-    );
-  }
-
-  // Render information about one movie
-  renderMovie(movie) {
-    return (
-      <View style={styles.container}>
-        <Image
-          source={{uri: movie.posters.thumbnail}}
-          style={styles.thumbnail}
-        />
-        <View style={styles.rightContainer}>
-          <Text style={styles.title}>{movie.title}</Text>
-          <Text style={styles.year}>{movie.year}</Text>
-        </View>
-      </View>
-    );
+		return (
+      <Tabs>
+        <Tab heading="Some Tab">
+          <Text>Tab awaiting new content</Text>
+        </Tab>
+        <Tab heading="Movies">
+          <MovieExample />
+        </Tab>
+      </Tabs>
+		);
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    flexDirection: 'row',
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#F5FCFF',
-  },
-  thumbnail: {
-    width: 53,
-    height: 81,
-  },
-  rightContainer: {
-   flex: 1,
-  },
-  title: {
-    fontSize: 20,
-    marginBottom: 8,
-    textAlign: 'center',
-  },
-  year: {
-    textAlign: 'center',
-  },
-  listView: {
-   paddingTop: 20,
-   backgroundColor: '#F5FCFF',
- },
-});
